Vertex BMatrix calculation added.

+++ src/main/scala/calculation/AbstractInput.scala
@@ -4,10 +4,10 @@ sealed trait AbstractInput
 
 case object EmptyInput extends AbstractInput
 
-case class SingleVertexInput(vertex: Long) extends AbstractInput {
+case class SingleVertexInput(vertex: Int) extends AbstractInput {
   override def toString = vertex.toString
 }
 
-case class MultipleVertexInput(vertex: Seq[Long]) extends AbstractInput {
+case class MultipleVertexInput(vertex: Seq[Int]) extends AbstractInput {
   override def toString = vertex.toString
 }
